requirements.txt
absl-py==1.0.0
addict==2.4.0
astroid==2.9.3
asttokens==2.0.5
backcall==0.2.0
backports.functools-lru-cache==1.6.4
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.12
cupy-cuda111==8.6.0
cycler==0.10.0
debugpy==1.5.1
decorator==4.4.2
decord==0.6.0
einops==0.4.1
entrypoints==0.4
executing==0.8.3
fastrlock==0.6
fvcore==0.1.5.post20220512
google-auth==2.6.6
google-auth-oauthlib==0.4.6
grpcio==1.44.0
idna==3.3
imageio==2.9.0
imageio-ffmpeg==0.4.4
importlib-metadata==4.11.3
iopath==0.1.9
ipykernel==6.13.0
ipython==8.2.0
isort==5.10.1
jedi==0.18.1
joblib==1.1.0
jupyter-client==7.0.6
jupyter-core==4.9.2
kiwisolver==1.3.1
lazy-object-proxy==1.7.1
Markdown==3.3.6
matplotlib==3.4.2
matplotlib-inline==0.1.3
mccabe==0.6.1
memory-profiler==0.60.0
mmcv-full==1.5.0
nest-asyncio==1.5.5
networkx==2.5.1
numpy==1.22.3
oauthlib==3.2.0
opencv-contrib-python==4.5.2.54
opencv-python==4.5.2.54
packaging==21.3
pandas==1.4.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.1.1
# Pillow-SIMD==9.0.0.post1
platformdirs==2.5.1
portalocker==2.4.0
prompt-toolkit==3.0.29
protobuf==3.20.1
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.11.2
pylint==2.12.2
pyparsing==2.4.7
python-dateutil==2.8.2
pytz==2021.3
PyWavelets==1.1.1
PyYAML==6.0
pyzmq==19.0.2
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
scikit-image==0.18.2
scipy==1.7.0
six==1.16.0
spatial-correlation-sampler==0.3.0
stack-data==0.2.0
tabulate==0.8.9
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
termcolor==1.1.0
thop==0.0.31.post2005241907
tifffile==2021.6.14
toml==0.10.2
torch==1.11.0+cu113
torch-tb-profiler==0.4.0
torchaudio==0.11.0+cu113
torchvision==0.12.0+cu113
tornado==6.1
tqdm==4.64.0
traitlets==5.1.1
typing_extensions==4.1.1
urllib3==1.26.8
wcwidth==0.2.5
Werkzeug==2.1.1
wrapt==1.13.3
yacs==0.1.8
yapf==0.32.0
zipp==3.8.0
ftfy==6.1.1
regex==2022.4.24
num2words==0.5.10
timm==0.5.4